move Enterprise settings types to general types file
