+++ enterprise/frontend/src/metabase-enterprise/api/utils/settings.ts
@@ -1,0 +1,4 @@
+import { _useAdminSetting } from "metabase/api";
+import type { EnterpriseSettingKey } from "metabase-types/api";
+
+export const useEnterpriseAdminSetting = _useAdminSetting<EnterpriseSettingKey>;
