+++ frontend/src/metabase/api/session.ts
@@ -1,6 +1,9 @@
 import MetabaseSettings from "metabase/lib/settings";
 import { loadSettings } from "metabase/redux/settings";
-import type { PasswordResetTokenStatus, Settings } from "metabase-types/api";
+import type {
+  EnterpriseSettings,
+  PasswordResetTokenStatus,
+} from "metabase-types/api";
 
 import { Api } from "./api";
 
@@ -23,7 +26,7 @@ export const sessionApi = Api.injectEndpoints({
         body: { email },
       }),
     }),
-    getSessionProperties: builder.query<Settings, void>({
+    getSessionProperties: builder.query<EnterpriseSettings, void>({
       query: () => ({
         method: "GET",
         url: "/api/session/properties",
