+++ frontend/src/metabase/api/utils/settings.ts
@@ -1,6 +1,6 @@
 import { useMemo } from "react";
 
-import type { Settings } from "metabase-types/api";
+import type { EnterpriseSettingKey, SettingKey } from "metabase-types/api";
 
 import { useGetSettingsQuery } from "../session";
 import {
@@ -10,8 +10,9 @@ import {
 
 /**
  * One hook to get setting values and mutators for a given setting
+ * generic version, should not be used directly
  */
-export const useAdminSetting = <SettingName extends keyof Settings>(
+export const _useAdminSetting = <SettingName extends EnterpriseSettingKey>(
   settingName: SettingName,
 ) => {
   const {
@@ -39,3 +40,8 @@ export const useAdminSetting = <SettingName extends keyof Settings>(
     ...apiProps,
   };
 };
+
+/**
+ * One hook to get setting values and mutators for a given OSS setting
+ */
+export const useAdminSetting = _useAdminSetting<SettingKey>;
