+++ frontend/test/__support__/settings.ts
@@ -1,6 +1,5 @@
 import MetabaseSettings from "metabase/lib/settings";
-import type { EnterpriseSettings } from "metabase-enterprise/settings/types";
-import type { Settings } from "metabase-types/api";
+import type { EnterpriseSettings, Settings } from "metabase-types/api";
 import { createMockSettings } from "metabase-types/api/mocks";
 import { createMockSettingsState } from "metabase-types/store/mocks";
